message [nfc]: Move proxy boilerplate out to substore file
